// services/station-service/src/stations/dto/query-stations.dto.d.ts
import { StationStatus } from '@omc-erp/shared-types';
export declare class QueryStationsDto {
    page?: number;
    limit?: number;
    search?: string;
    status?: StationStatus;
    region?: string;
    isActive?: boolean;
    latitude?: number;
    longitude?: number;
    radius?: number;
}
//# sourceMappingURL=query-stations.dto.d.ts.map